refactor: move handlers to separate folders
