+++ packages/product-service/getProductById/getProductById.ts
@@ -1,9 +1,12 @@
 import { APIGatewayProxyHandler } from 'aws-lambda';
 import 'source-map-support/register';
-import { cors } from './cors';
-import { products } from './products';
+import { cors } from '../cors';
+import { products } from '../products';
 
-export const handler: APIGatewayProxyHandler = async (event, _context) => {
+export const getProductById: APIGatewayProxyHandler = async (
+  event,
+  _context,
+) => {
   const product = products.find(({ id }) => event.pathParameters?.id === id);
 
   if (!product) {
