+++ packages/product-service/getProductById/index.ts
@@ -1,0 +1,1 @@
+export * from './getProductById';
