--- packages/product-service/getProductsList.ts
@@ -1,12 +1,0 @@
-import { APIGatewayProxyHandler } from 'aws-lambda';
-import 'source-map-support/register';
-import { cors } from './cors';
-import { products } from './products';
-
-export const handler: APIGatewayProxyHandler = async () => {
-    return {
-        statusCode: 200,
-        body: JSON.stringify(products),
-        headers: cors,
-    };
-}
