+++ packages/product-service/getProductsList/index.ts
@@ -1,0 +1,1 @@
+export * from './getProductsList';
