+++ packages/product-service/handlers.ts
@@ -1,0 +1,2 @@
+export * from './getProductById';
+export * from './getProductsList';
